+++ spec/rails_helper.rb
@@ -1094,21 +1094,6 @@ RSpec.configure do |config|
     MessageBus.backend_instance.reset! # Clears all existing backlog from memory backend
   end
 
-  config.before(:each, type: :multisite) do
-    Rails.configuration.multisite = true # rubocop:disable Discourse/NoDirectMultisiteManipulation
-
-    RailsMultisite::ConnectionManagement.config_filename = "spec/fixtures/multisite/two_dbs.yml"
-
-    RailsMultisite::ConnectionManagement.establish_connection(db: "default")
-  end
-
-  config.after(:each, type: :multisite) do
-    ActiveRecord::Base.connection_handler.clear_all_connections!
-    Rails.configuration.multisite = false # rubocop:disable Discourse/NoDirectMultisiteManipulation
-    RailsMultisite::ConnectionManagement.clear_settings!
-    ActiveRecord::Base.establish_connection
-  end
-
   class TestCurrentUserProvider < Auth::DefaultCurrentUserProvider
     def log_on_user(user, session, cookies, opts = {})
       # Try using the main session as `session` sometimes is a server session
@@ -1120,19 +1105,6 @@ RSpec.configure do |config|
       # Try using the main session as `session` sometimes is a server session
       (cookies.try(:request).try(:session) || session).delete(:current_user_id)
       super
-    end
-  end
-
-  # Normally we `use_transactional_fixtures` to clear out a database after a test
-  # runs. However, this does not apply to tests done for multisite. The second time
-  # a test runs you can end up with stale data that breaks things. This method will
-  # force a rollback after using a multisite connection.
-  def test_multisite_connection(name)
-    RailsMultisite::ConnectionManagement.with_connection(name) do
-      ActiveRecord::Base.transaction(joinable: false) do
-        yield
-        raise ActiveRecord::Rollback
-      end
     end
   end
 end
